tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.arcTo.scale.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.arcTo.scale</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.arcTo.scale</h1>
<p class="desc">arcTo scales the curve, not just the control points</p>


<script>
var t = async_test("arcTo scales the curve, not just the control points");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.fillStyle = '#0f0';
  ctx.beginPath();
  ctx.moveTo(0, 50);
  ctx.translate(100, 0);
  ctx.scale(0.1, 1);
  ctx.arcTo(50, 50, 50, 0, 50);
  ctx.lineTo(-1000, 0);
  ctx.fill();

  _assertPixel(canvas, 0,0, 0,255,0,255);
  _assertPixel(canvas, 50,0, 0,255,0,255);
  _assertPixel(canvas, 99,0, 0,255,0,255);
  _assertPixel(canvas, 0,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 99,25, 0,255,0,255);
  _assertPixel(canvas, 0,49, 0,255,0,255);
  _assertPixel(canvas, 50,49, 0,255,0,255);
  _assertPixel(canvas, 99,49, 0,255,0,255);
  t.done();

});
</script>
